requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
crypto==1.4.1
evdev==1.6.1
Faker==18.3.1
frozenlist==1.3.3
geoip2==4.6.0
gTTS==2.3.1
html5lib==1.1
idna==3.4
maxminddb==2.2.0
multidict==6.0.4
Naked==0.1.32
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
pycryptodome==3.17
pynput==1.7.6
pypdf2==3.0.1
pypng==0.20220715.0
pyshorteners==1.0.1
python-dateutil==2.8.2
python-xlib==0.33
pyttsx3==2.90
pytube==12.1.3
PyYAML==6.0
qrcode==7.4.2
requests==2.28.2
shellescape==3.8.1
six==1.16.0
soupsieve==2.4
typing-extensions==4.5.0
urllib3==1.26.15
webencodings==0.5.1
yarl==1.8.2
